apps/server/pyproject.toml
[project]
name = "cognix-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "geetheswar-v", email = "geetheswar.edu@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "email-validator>=2.1.0",
    "fastapi>=0.128.1",
    "httpx>=0.28.1",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.32",
    "uuid6>=2025.0.1",
    "uvicorn[standard]>=0.40.0",
]
